requirements.txt
# ── Core Runtime ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# ── Data Layer ─────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ── LLM Layer ─────────────────────────────────────────────────────────────────
google-genai>=1.0.0

# ── Input Validation ──────────────────────────────────────────────────────────
pydantic>=2.5.0
